+++ MangoPay/InternationalAccount.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace MangoPay;
+
+class InternationalAccount extends Libraries\Dto
+{
+    /**
+     * @var string
+     */
+    public $Iban;
+
+    /**
+     * @var string
+     */
+    public $Bic;
+}