+++ MangoPay/Libraries/UrlTool.php
@@ -12,7 +12,8 @@ class UrlTool
 
     /**
      * Constructor
-     * @param \MangoPay\MangoPayApi Root/parent instance that holds the OAuthToken and Configuration instance
+     * Root/parent instance that holds the OAuthToken and Configuration instance
+     * @param \MangoPay\MangoPayApi $root
      */
     public function __construct($root)
     {